Added corner detection implementation to calculate high goal distance
Using a linear search algorithm on the contour point list, the updated BlueGoalVisionPipeline can determine the corners of the high goal. Using these corners, the pipeline can  determine the height of the center of the goal, which will not change depnding on the camera angle. This height will be used in a calibration curve to calculate distance from the goal in the future.